u-boot-xlnx : Add support for building BOOT.BIN file.
If the command bootgen is in the path, and there is a file for
the current MACHINE called fsbl.elf.bz2 available, the BOOT.BIN
file will be built.

I had to change the version to remove the + signs from the file name.
Otherwise the script failed. I need to look at this harder and come up with
a better fix.

Code based on:

https://github.com/milosoftware/meta-zynq/blob/master/recipes-bsp/u-boot/u-boot-zynq_git.bb

Signed-off-by: Philip Balister <philip@balister.org>